combine unit tests #2374
